debugging

+++ app/src/androidTest/java/nl/codingwithlinda/smartstep/features/settings/presentation/weight_settings/WeightSettingsScreenTest.kt
@@ -57,7 +57,7 @@ class WeightSettingsScreenTest {
                     when (it) {
                        is ActionWeightInput.ChangeSystem ->{
                            uiState.update {
-                               WeightSettingUiState.Imperial(weightRangeKg.last())
+                               WeightSettingUiState.Imperial(weightRangeKg.last().toDouble())
                            }
 
                        }
