+++ app/src/main/java/nl/codingwithlinda/smartstep/core/domain/model/settings/UserSettings.kt
@@ -2,6 +2,6 @@ package nl.codingwithlinda.smartstep.core.domain.model.settings
 
 data class UserSettings(
     val gender: Gender = Gender.FEMALE,
-    val weight: Int = 60,
+    val weight: Double = 60.0,
     val height: Int = 170,
 )
